LayoutTests/accessibility/mac/table-visible-rows-expected.txt

This tests that AXVisibleRows returns visible rows correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS rows is 5
PASS visibleRows is 3
PASS successfullyParsed is true

TEST COMPLETE
